shark: limit fuzz threshold to 16 bits.
This gives the fuzzer a little more scope than the previous 8-bit
limit without constantly triggering out-of-memory conditions.

If we're not going to make the entrypoints fallible on bad arguments,
writing a custom Arbitrary impl to avoid both zero (immediate index
panics) and very large share numbers (wasting memory) while still
covering a realistic range of threshold values.

+++ sharks/fuzz/fuzz_targets/recover.rs
@@ -5,8 +5,9 @@ use arbitrary::Arbitrary;
 use sharks::{Share, Sharks};
 
 #[derive(Debug, Arbitrary)]
+// Limit threshhold to 16 bits so we don't exhaust memory.
 struct Parameters {
-    pub threshold: u8,
+    pub threshold: u16,
     pub shares: Vec<Share>,
 }
